server/extractor/requirements.txt
flask==3.0.3
python-pptx==0.6.23
pymupdf==1.24.9
pillow==10.3.0
pytesseract==0.3.10
